BFP2025G2-25/back

+++ src/main/java/com/campusdual/bfp/model/dto/CandidateTagsDTO.java
@@ -4,12 +4,12 @@ import java.util.List;
 
 public class CandidateTagsDTO {
     private int candidateId;
-    private List<Long> tagIds;
+    private List<Integer> tagIds;
 
     public CandidateTagsDTO() {
     }
 
-    public List<Long> getTagIds() {
+    public List<Integer> getTagIds() {
         return tagIds;
     }
 }
